Add variable to register list of currently installed APDU services
This is needed when an APDU service is modified by changing its manifest
and removing the APDU service meta-data. The service is then no longer an
APDU service and the routing table is not updated.
The list of installed services allows to identify that this application
was at some point an APDU service and we can trigger the routing table
update.

Codex #1197107